Refactor new order creation logic and add order items
This commit refactors the logic for creating a new order in the NewOrder component. It introduces an onSuccess callback function that is triggered after a successful order creation. In this callback, the order ID is logged to the console and the createOrderItems function is called to create order items for the newly created order.

The code changes also include updating the orderState property from a string to a number and removing unnecessary code related to the onSuccess callback in the mutate function.

These changes improve the clarity and efficiency of the new order creation process.